Address GPUI history BDD review feedback (#115)
Refine shared BDD support, remove manual copies, and split the
multi-shape bindings into the history BDD module directory.

Document the v0.6 stateful-GPUI constraint and record the fixed
scenario migration without overstating the outstanding roadmap audit.
